whatsapp-mcp-server/pyproject.toml
[project]
name = "whatsapp-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "google-api-core>=2.25.1",
    "google-genai>=1.24.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "requests>=2.32.3",
]
